chore: updated volto and fixed video block
