+++ package.json
@@ -184,7 +184,7 @@
     "@fortawesome/free-regular-svg-icons": "5.14.0",
     "@fortawesome/free-solid-svg-icons": "5.14.0",
     "@fortawesome/react-fontawesome": "0.1.11",
-    "@plone/volto": "plone/volto#f435e097263b184873f4eb96cfb70af050fe4a26",
+    "@plone/volto": "plone/volto#fedfca2c894b0a288af2eff95ef9fc8d19a85eec",
     "@tinymce/tinymce-react": "3.8.1",
     "bootstrap-italia": "1.3.9",
     "classnames": "2.2.6",
