+++ src/customizations/volto/components/manage/Blocks/Video/Edit.jsx
@@ -5,8 +5,8 @@
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { defineMessages, FormattedMessage, injectIntl } from 'react-intl';
-import { Button, Input, Embed, Message } from 'semantic-ui-react';
+import { defineMessages, injectIntl } from 'react-intl';
+import { Button, Input, Message } from 'semantic-ui-react';
 import cx from 'classnames';
 import { isEqual } from 'lodash';
 
@@ -14,12 +14,7 @@ import { Icon, SidebarPortal, VideoSidebar } from '@plone/volto/components';
 import clearSVG from '@plone/volto/icons/clear.svg';
 import aheadSVG from '@plone/volto/icons/ahead.svg';
 import videoBlockSVG from '@plone/volto/components/manage/Blocks/Video/block-video.svg';
-import {
-  isInternalURL,
-  getParentUrl,
-  flattenToAppURL,
-} from '@plone/volto/helpers';
-import config from '@plone/volto/registry';
+import Body from './Body.jsx';
 
 const messages = defineMessages({
   VideoFormDescription: {
@@ -146,12 +141,6 @@ class Edit extends Component {
     const placeholder =
       this.props.data.placeholder ||
       this.props.intl.formatMessage(messages.VideoBlockInputPlaceholder);
-
-    const allowsExternals =
-      data.allowExternals !== undefined
-        ? !!data.allowExternals
-        : !!config.settings.videoAllowExternalsDefault;
-
     return (
       <div
         className={cx(
@@ -164,159 +153,7 @@ class Edit extends Component {
         )}
       >
         {data.url ? (
-          <div
-            className={cx('video-inner', {
-              'full-width': this.props.data.align === 'full',
-            })}
-          >
-            {data.url.match('youtu') ? (
-              <>
-                {data.url.match('list') ? (
-                  data.preview_image ? (
-                    <Embed
-                      url={`https://www.youtube.com/embed/videoseries?list=${
-                        data.url.match(/^.*\?list=(.*)$/)[1]
-                      }`}
-                      placeholder={
-                        isInternalURL(data.preview_image)
-                          ? `${flattenToAppURL(
-                              data.preview_image,
-                            )}/@@images/image`
-                          : data.preview_image
-                      }
-                      defaultActive
-                      autoplay={false}
-                    />
-                  ) : (
-                    <Embed
-                      url={`https://www.youtube.com/embed/videoseries?list=${
-                        data.url.match(/^.*\?list=(.*)$/)[1]
-                      }`}
-                      icon="play"
-                      defaultActive
-                      autoplay={false}
-                    />
-                  )
-                ) : data.preview_image ? (
-                  <Embed
-                    id={
-                      data.url.match(/.be\//)
-                        ? data.url.match(/^.*\.be\/(.*)/)[1]
-                        : data.url.match(/^.*\?v=(.*)$/)[1]
-                    }
-                    source="youtube"
-                    placeholder={
-                      isInternalURL(data.preview_image)
-                        ? `${flattenToAppURL(
-                            data.preview_image,
-                          )}/@@images/image`
-                        : data.preview_image
-                    }
-                    icon="play"
-                    autoplay={false}
-                  />
-                ) : (
-                  <Embed
-                    id={
-                      data.url.match(/.be\//)
-                        ? data.url.match(/^.*\.be\/(.*)/)[1]
-                        : data.url.match(/^.*\?v=(.*)$/)[1]
-                    }
-                    source="youtube"
-                    icon="play"
-                    defaultActive
-                    autoplay={false}
-                  />
-                )}
-              </>
-            ) : (
-              <>
-                <div className="ui blocker" />
-                {data.url.match('vimeo') ? (
-                  data.preview_image ? (
-                    <Embed
-                      id={data.url.match(/^.*\.com\/(.*)/)[1]}
-                      source="vimeo"
-                      placeholder={
-                        isInternalURL(data.preview_image)
-                          ? `${flattenToAppURL(
-                              data.preview_image,
-                            )}/@@images/image`
-                          : data.preview_image
-                      }
-                      icon="play"
-                      autoplay={false}
-                    />
-                  ) : (
-                    <Embed
-                      id={data.url.match(/^.*\.com\/(.*)/)[1]}
-                      source="vimeo"
-                      icon="play"
-                      defaultActive
-                      autoplay={false}
-                    />
-                  )
-                ) : (
-                  <>
-                    <div className="ui blocker" />
-                    {data.url.match('.mp4') ? (
-                      // eslint-disable-next-line jsx-a11y/media-has-caption
-                      <video
-                        src={
-                          isInternalURL(
-                            data.url.replace(
-                              getParentUrl(config.settings.apiPath),
-                              '',
-                            ),
-                          )
-                            ? `${data.url}/@@download/file`
-                            : data.url
-                        }
-                        controls
-                        poster={
-                          data.preview_image
-                            ? isInternalURL(data.preview_image)
-                              ? `${flattenToAppURL(
-                                  data.preview_image,
-                                )}/@@images/image`
-                              : data.preview_image
-                            : ''
-                        }
-                        type="video/mp4"
-                      />
-                    ) : data.url && allowsExternals ? (
-                      // eslint-disable-next-line jsx-a11y/media-has-caption
-                      <video
-                        src={data.url}
-                        controls
-                        poster={
-                          data.preview_image
-                            ? isInternalURL(data.preview_image)
-                              ? `${flattenToAppURL(
-                                  data.preview_image,
-                                )}/@@images/image`
-                              : data.preview_image
-                            : null
-                        }
-                        type="video/mp4"
-                      />
-                    ) : (
-                      <div>
-                        <Message>
-                          <center>
-                            <FormattedMessage
-                              id="Please enter a valid URL by deleting the block and adding a new video block."
-                              defaultMessage="Please enter a valid URL by deleting the block and adding a new video block."
-                            />
-                          </center>
-                        </Message>
-                      </div>
-                    )}
-                  </>
-                )}
-              </>
-            )}
-          </div>
+          <Body data={this.props.data} isEditMode={true} />
         ) : (
           <Message>
             <center>
